backend/requirements.txt
accelerate==0.21.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
Bottleneck
bs4==0.0.1
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
Chroma==0.2.0
chroma-bullet==2.2.0
chroma-hnswlib==0.7.3
chroma-migrate==0.0.7
chromadb==0.3.29
click==8.1.7
clickhouse-connect==0.6.6
cmake==3.28.3
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.3
cycler==0.12.1
dataclasses-json==0.5.14
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
duckdb==0.7.1
effdet==0.4.1
einops==0.6.1
emoji==2.10.1
exceptiongroup==1.2.0
fastapi==0.85.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.28.0
googleapis-common-protos==1.62.0
gpt4all==2.2.1.post1
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
hnswlib==0.8.0
httptools==0.6.1
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
jsonpath-python==1.0.6
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.0.267
langdetect==1.0.9
langsmith==0.0.92
layoutparser==0.3.4
lit==17.0.6
lxml==5.1.0
lz4==4.3.3
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
mkl-fft
mkl-random
mkl-service==2.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr
numpy 
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openapi-schema-pydantic==1.2.4
opencv-python==4.9.0.80
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.7.0
packaging==23.2
pandas
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pikepdf==8.13.0
pillow==10.2.0
pillow_heif==0.15.0
portalocker==2.8.2
posthog==3.4.2
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.2
PyMuPDF==1.23.24
PyMuPDFb==1.23.22
pyparsing==3.1.1
pypdf==4.0.2
pypdfium2==4.27.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pytesseract==0.3.10
python-dateutil
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
pytz
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six 
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.20.4
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.3.0
timm==0.9.16
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.2
transformers==4.31.0
triton==2.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata
unstructured==0.12.4
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==2.2.1
uvicorn==0.18.3
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
